Move legacy Outcome out of core
